pyproject.toml
[project]
name = "uptime-ping"
version = "0.1.0"
description = "Simple API health monitoring with Telegram alerts"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
